Create ssi-table
Python script to easily calculate sliding scale dosing. NOTE THAT THIS IS NOT TO BE USED FOR MEDICAL ADVICE AND ONLY FOR TESTING PURPOSES. Insulin dosing should always be calculated and verified by a licensed provider. This tool aims to assist with the calculation but should NEVER be used without verification by the licensed provider.